New function: ip_match; Updated usingProxy; Updated errtostr; Syntax changes
Added a new function to check if a given IP address is included in a list of given CIDR range, and which is the first it matches.

Updated usingProxy to allow users to specify their own ports rather than the given preset ones. Also fixed it to work properly again.

Shortened errortostr by removing unecessary breaks.

Changed syntax of comments so they all match properly and use correct terms (author instead of credits).